Precompute HTML for icons
This creates an "Icon" class to precompute and provide
HTML for icons.  Icons are used in all HTML output, so
it makes sense to precompute it for performance.
