+++ app/views/projects/show.html.erb
@@ -1,9 +1,9 @@
 <% content_for :nav_extras do %>
     <% if can_edit? %>
-      <li><%= link_to my_icon_tag('fa-edit') + t('.edit'), edit_project_path(@project, criteria_level: @criteria_level) %></li>
+      <li><%= link_to Icon[:'fa-edit'] + t('.edit'), edit_project_path(@project, criteria_level: @criteria_level) %></li>
     <% end %>
     <% if can_control? %>
-      <li><%= link_to my_icon_tag('fa-times-circle') + t('.delete'), project_path(@project) + '/delete_form' %></li>
+      <li><%= link_to Icon[:'fa-times-circle'] + t('.delete'), project_path(@project) + '/delete_form' %></li>
     <% end %>
 <% end %>
