+++ app/views/users/show.html.erb
@@ -9,7 +9,7 @@
          <%= avatar_for @user %>
          <%= @user.name || @user.nickname %>
          <% if current_user && (@user == current_user || current_user.admin?) %>
-           <%= link_to my_icon_tag('fa-edit') + t('.edit_user'),
+           <%= link_to Icon[:'fa-edit'] + t('.edit_user'),
                  edit_user_path(@user),
                  class: 'btn btn-primary', rel: 'nofollow' %>
          <% end %>
@@ -50,7 +50,7 @@
       | <a href="mailto:<%= CGI::escape(@user.email).html_safe %>"><%=
         t('.send_email_to') + @user.email.to_s %></a>
     <% end %>
-    | <%= link_to my_icon_tag('fa-times-circle') + t('.delete_link_name'),
+    | <%= link_to Icon[:'fa-times-circle'] + t('.delete_link_name'),
                   @user, method: :delete,
                   data: { confirm: t('.confirm_delete') },
                   rel: 'nofollow' %>
