+++ config/initializers/icon.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+# Precalculate HTML for icons
+Icon.initialize_class
